Consolidating android specific logic
