+++ trueseeing/app/cmd/android/alias.py
