+++ trueseeing/app/cmd/android/exploit.py
@@ -64,8 +64,8 @@ class ExploitCommand(Command):
     import os
     import time
     from tempfile import TemporaryDirectory
-    from trueseeing.core.asm import APKAssembler
-    from trueseeing.core.tools import move_apk
+    from trueseeing.core.android.asm import APKAssembler
+    from trueseeing.core.android.tools import move_apk
 
     apk = self._runner._target
     origapk = apk.replace('.apk', '.apk.orig')
@@ -130,7 +130,7 @@ class ExploitCommand(Command):
 
       # XXX
       path = f'resources/package_1/res/xml/{key}.xml'
-      q.patch_put(path, (files('trueseeing')/'libs'/'nsc.xml').read_bytes())
+      q.patch_put(path, (files('trueseeing')/'libs'/'android'/'nsc.xml').read_bytes())
 
       # XXX
       import lxml.etree as ET
@@ -243,7 +243,7 @@ class ExploitCommand(Command):
 
     import time
     import re
-    from trueseeing.core.device import AndroidDevice
+    from trueseeing.core.android.device import AndroidDevice
 
     at = time.time()
     nr = 0
@@ -278,8 +278,9 @@ class ExploitCommand(Command):
 
     import time
     from subprocess import CalledProcessError
-    from trueseeing.core.device import AndroidDevice
-    from trueseeing.core.tools import toolchains, invoke_passthru
+    from trueseeing.core.tools import invoke_passthru
+    from trueseeing.core.android.device import AndroidDevice
+    from trueseeing.core.android.tools import toolchains
 
     at = time.time()
     dev = AndroidDevice()
@@ -370,8 +371,9 @@ class ExploitCommand(Command):
 
     import time
     from subprocess import CalledProcessError
-    from trueseeing.core.device import AndroidDevice
-    from trueseeing.core.tools import toolchains, invoke_passthru
+    from trueseeing.core.tools import invoke_passthru
+    from trueseeing.core.android.device import AndroidDevice
+    from trueseeing.core.android.tools import toolchains
 
     at = time.time()
     dev = AndroidDevice()
