+++ trueseeing/core/model/sig.py
@@ -7,7 +7,7 @@ from pubsub import pub
 
 if TYPE_CHECKING:
   from typing import ClassVar
-  from trueseeing.core.context import Context
+  from trueseeing.core.android.context import Context
   from trueseeing.core.model.issue import Issue
 
 class Detector(ABC):
