+++ trueseeing/libs/android/nsc.xml
