+++ trueseeing/libs/android/store.0.sql
