+++ trueseeing/libs/android/store.1.sql
