+++ trueseeing/sig/android/manifest.py
