+++ trueseeing/sig/android/security.py
@@ -6,8 +6,8 @@ import io
 import re
 import os
 
-from trueseeing.core.model.code import InvocationPattern
-from trueseeing.core.analysis.flow import DataFlows
+from trueseeing.core.android.model.code import InvocationPattern
+from trueseeing.core.android.analysis.flow import DataFlows
 from trueseeing.core.model.sig import Detector
 from trueseeing.core.model.issue import Issue
 from trueseeing.core.ui import ui
